Исправлена валидация Заменены параметры модели

+++ Client.Core/Data/Domain.cs
@@ -6,17 +6,17 @@ public record Domain
 
     public required double EndX { get; init; }
 
-    public required int SplitsXCount { get; init; }
+    public required double SplitsXCount { get; init; }
 
     public required double StartY { get; init; }
 
     public required double EndY { get; init; }
 
-    public required int SplitsYCount { get; init; }
+    public required double SplitsYCount { get; init; }
 
     public required double StartZ { get; init; }
 
     public required double EndZ { get; init; }
 
-    public required int SplitsZCount { get; init; }
+    public required double SplitsZCount { get; init; }
 }
